refactor: 함수 분리 + 함수 이름 변경 + 테스트코드에서 불필요한 함수 제거
