+++ src/js/app.js
@@ -1,12 +1,7 @@
 import { $ } from './utils/dom.js';
 import { getLottoPrice, checkLottoPrice } from './core/checkInputValue.js';
-import { makeLottos } from './core/playLotto.js';
-import {
-  renderLastLottoNumber,
-  renderPurchasedLottoList,
-} from './views/render.js';
 import { handleToggleButtonClick } from './component/toggleButton.js';
-
+import { drawLotto } from './core/drawLotto.js';
 export default class App {
   constructor() {
     this.lottoPrice = 0;
@@ -15,16 +10,15 @@ export default class App {
   handleEvent() {
     $('.lotto-price-input-form').addEventListener(
       'submit',
-      this.handlePriceInputEvent,
+      this.handlePriceInputSubmit,
     );
-
     $('.purchased-lotto-list-container').addEventListener(
       'click',
       handleToggleButtonClick,
     );
   }
 
-  handlePriceInputEvent(event) {
+  handlePriceInputSubmit(event) {
     event.preventDefault();
     const lottoPrice = checkLottoPrice(getLottoPrice());
     if (!lottoPrice) {
@@ -33,9 +27,7 @@ export default class App {
     }
     this.lottoPrice = lottoPrice;
     $('.lotto-price-input').disabled = true;
-    makeLottos(this.lottoPrice / 1000);
-    renderPurchasedLottoList(this.lottoPrice / 1000);
-    renderLastLottoNumber();
+    drawLotto(this.lottoPrice);
   }
 }
